eliminates HrsUrlHelper in favor of adding fromString() to HrsUrl
